Server/requirements.txt
fastmcp
pandas
pyserial
